fixing hexa

+++ check_flag.c
@@ -6,7 +6,7 @@
 /*   By: jbelkerf <jbelkerf@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/07 17:17:32 by jbelkerf          #+#    #+#             */
-/*   Updated: 2024/11/27 18:14:57 by jbelkerf         ###   ########.fr       */
+/*   Updated: 2024/11/27 22:19:05 by jbelkerf         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int	check_flag(char c, va_list ap)
 	else if (c == 'u')
 		counter = print_unsigned(va_arg(ap, int), 0);
 	else if (c == 'x')
-		counter = hexa_l(va_arg(ap, long int), 0);
+		counter = hexa_l(va_arg(ap, long long int), 0);
 	else if (c == 'X')
 		counter = hexa_u(va_arg(ap, long int), 0);
 	else if (c == '%')
